feat(extension-chrome): refactor code after code review
